Added course ratings
